feature flag

+++ Sources/PointFree/OfficeHoursMiddleware.swift
@@ -10,7 +10,13 @@ import Views
 func officeHoursMiddleware(
   _ conn: Conn<StatusLineOpen, OfficeHoursRoute>
 ) async -> Conn<ResponseEnded, Data> {
+  @Dependency(\.currentUser) var currentUser
   @Dependency(\.database) var database
+
+  guard currentUser.hasAccess(to: .officeHours) else {
+    return routeNotFoundMiddleware(conn)
+  }
+
   do {
     switch conn.data {
     case let .deleteQuestion(id):
